Changing to connect everytime
Since we had problems using this lib in bob, we decided to change the
behaviour to connect everytime you call invoke. The connect function is
now deprecated and only returns a map to avoid breaking the api. You
still are able to use client like before but it is discouraged now
to use connect and only using invoke is also sufficient.

I moved the connect-function to the requests namespace because
it is not part of the official api but is used with the fetch-function
and therefore should reside with it. The panic! function has to be
in the requests namespace as well because otherwise we would have a
cyclic dependency. The documentation is changed accordingly and fetch
needs to call the for now unofficial function connect*.

Refers https://github.com/lispyclouds/clj-docker-client/issues/20